+++ core/core-api/src/main/java/com/ticket/core/api/controller/AuthController.java
@@ -1,10 +1,9 @@
 package com.ticket.core.api.controller;
 
 import com.ticket.core.api.controller.docs.AuthControllerDocs;
+import com.ticket.core.domain.auth.token.AuthRefreshToken;
 import com.ticket.core.domain.auth.usecase.*;
 import com.ticket.core.domain.member.MemberPrincipal;
-import com.ticket.core.support.exception.AuthException;
-import com.ticket.core.support.exception.ErrorType;
 import com.ticket.core.support.response.ApiResponse;
 import com.ticket.core.support.util.CookieUtils;
 import jakarta.servlet.http.HttpServletResponse;
