+++ core/core-api/src/main/java/com/ticket/core/domain/auth/usecase/LogoutUseCase.java
@@ -1,5 +1,6 @@
 package com.ticket.core.domain.auth.usecase;
 
+import com.ticket.core.domain.auth.token.AuthRefreshToken;
 import com.ticket.core.domain.auth.token.AuthTokenManager;
 import com.ticket.core.domain.auth.token.RefreshTokenService;
 import com.ticket.core.support.exception.AuthException;
